pyproject.toml
[build-system]
build-backend = "setuptools.build_meta"
requires = ["setuptools>=64", "setuptools-scm>=8"]

[project]
name = "mkdocs-rss-plugin"
description = "MkDocs plugin to generate RSS and JSON feeds using Mkdocs site configuration, git log and Mkdocs pages'meta."
dynamic = ["version"]
readme = "README.md"
license = "MIT"
license-files = ["LICENSE"]
requires-python = ">=3.10"
authors = [{ name = "Julien Moura", email = "dev@ingeoveritas.com" }]
maintainers = [{ name = "Julien Moura", email = "dev@ingeoveritas.com" }]
keywords = [
    "atom",
    "documentation",
    "feed",
    "git",
    "json-feed",
    "mkdocs",
    "plugin",
    "rss",
]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "Intended Audience :: Information Technology",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Documentation",
    "Topic :: Software Development :: Documentation",
    "Topic :: Text Processing :: Markup :: Markdown",
]

dependencies = [
    "CacheControl[filecache]>=0.14.3,<1",
    "GitPython>=3.1.45,<3.2",
    "mkdocs>=1.6.1,<2",
    "requests>=2.32.5,<3",
    "tzdata==2024.* ; sys_platform == 'win32'",
]

[project.optional-dependencies]
dev = [
    "black>=25.9",
    "flake8>=7.1",
    "flake8-bugbear>=24.10",
    "flake8-builtins>=2.5",
    "flake8-eradicate>=1.5",
    "flake8-isort>=6.1.1",
    "Flake8-pyproject>=1.2.3",
    "pre-commit>=4,<5",
]
docs = [
    "mkdocs-git-committers-plugin-2>=2.4.1,<2.6",
    "mkdocs-git-revision-date-localized-plugin>=1.3,<1.6",
    "mkdocs-material[imaging]>=9.7,<9.8",
    "mkdocs-minify-plugin>=0.8,<0.9",
    "mkdocstrings-python>=1.16.2,<1.19",
    "termynal>=0.12.2,<0.14",
]
test = [
    "feedparser>=6.0.12,<6.1",
    "jsonfeed-util>=1.2,<2",
    "mkdocs-material[imaging]>=9.7,<9.8",
    "pytest-cov>=6.9.1,<8",
    "validator-collection>=1.5,<1.6",
]

[project.entry-points."mkdocs.plugins"]
rss = "mkdocs_rss_plugin.plugin:GitRssPlugin"

[project.urls]
Changelog = "https://github.com/guts/mkdocs-rss-plugin/blob/main/CHANGELOG.md"
Documentation = "https://guts.github.io/mkdocs-rss-plugin/"
Homepage = "https://guts.github.io/mkdocs-rss-plugin/"
Issues = "https://github.com/guts/mkdocs-rss-plugin/issues/"
releasenotes = "https://github.com/guts/mkdocs-rss-plugin/releases/latest"
Repository = "https://github.com/guts/mkdocs-rss-plugin/"


# -- TOOLING CONFIGURATION --

# Black configuration
[tool.black]
target-version = ["py310", "py311", "py312", "py313"]

# Coverage configuration
[tool.coverage.report]
exclude_also = [
    # Don't complain about abstract methods, they aren't run:
    "@(abc\\.)?abstractmethod",
    "class .*\\bProtocol\\):",
    # Don't complain about missing debug-only code:
    "def __repr__",
    "if self.debug:",
     # Don't complain if non-runnable code isn't run:
    "if 0:",
    "if __name__ == .__main__.:",
    # Don't complain about imports for type checking:
    "if TYPE_CHECKING:",
    # Don't complain if tests don't hit defensive assertion code:
    "raise AssertionError",
    "raise NotImplementedError",
    # ignore lines marked with a pragma
    "pragma: no cover",
]
ignore_errors = true
show_missing = true

[tool.coverage.run]
branch = true
omit = ["*/tests/*", "*/__pycache__/*", "*/.venv/*", "*/_version.py"]

[tool.flake8]
count = true
exclude = [
    ".git",
    "__pycache__",
    "docs/conf.py",
    "old",
    "build",
    "dist",
    ".venv*",
    "tests",
]
ignore = ["E121", "E123", "E126", "E203", "E226", "E24", "E704", "W503", "W504"]
max-complexity = 15
max-doc-length = 130
max-line-length = 100
statistics = true
tee = true

[tool.isort]
profile = "black"
ensure_newline_before_comments = true
force_grid_wrap = 0
include_trailing_comma = true
line_length = 88
multi_line_output = 3
use_parentheses = true

# Pytest configuration
[tool.pytest.ini_options]
addopts = [
    "--junitxml=junit/test-results.xml",
    "--cov=mkdocs_rss_plugin",
    "--cov-config=pyproject.toml",
    "--cov-report=html",
    "--cov-report=term",
    "--cov-report=xml",
    "--ignore=tests/_wip/",
]
junit_family = "xunit2"
minversion = "7.0"
norecursedirs = ".* build dev development dist docs CVS fixtures _darcs {arch} *.egg venv _wip"
python_files = ["test_*.py"]
testpaths = ["tests"]

# Setuptools configuration
[tool.setuptools]
include-package-data = true

[tool.setuptools.package-data]
mkdocs_rss_plugin = ["templates/*.jinja2"]

[tool.setuptools.packages.find]
include = ["mkdocs_rss_plugin"]
namespaces = false
where = ["."]

# setuptools-scm configuration
[tool.setuptools_scm]
fallback_version = "0.0.0+unknown"
local_scheme = "no-local-version"
version_scheme = "guess-next-dev"
write_to = "mkdocs_rss_plugin/_version.py"
